Add second location
