+++ static/models/test-couch.mtl
@@ -1,0 +1,8 @@
+# WaveFront *.mtl file (generated by CINEMA 4D)
+
+newmtl test-04
+Ka 1 1 1
+Kd 1 1 1
+map_Kd test-04.jpg
+illum 7
+
